image: Add a Weston image for general use for EMLinux machines
This installs the Weston and Mesa Debian packages on the default Weston
image for EMLinux 3.x.

Also it introduces poky's systemd startup script for Weston by migrating
to Isar. In the original script, a non-root user launches the Weston
process, but this script adopts root for that due to missing DRM
master permission when Weson is launched by a non-root user.