classes/cve-check: Fix overwriting patched status by cip-kernel-sec
The kernel-cve-check works by combining the results of poky's cve-check
and cip-kernel-sec's script. If a CVE is listed in CVE_CHECK_WHITELIST,
the CVE status should be "Patched" regardless of the cip-kernel-sec.
However, the current implementaion overwrites the status to "Unpatched"
when the cip-kernel-sec reports the CVE is not fixed.

To resolve this situation, we exclude the CVEs in the CVE_CHECK_WHITELIST
from the CVEs whose status should be corrected.

Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>