cve-update-nvd2-native: Create CVE_CHECK directory if it is not present
The pre-download feature of the CVE database file moves the downloaded
file to $DL_DIR/CVE_CHECK. However, during the first execution of the
cve-check task, the destination directory does not exist at the time the
database file is downloaded, causing the file move operation to fail.
Therefore, after downloading the database file, we will check if the
destination directory exists, and if it does not, we will create it.
Signed-off-by: Masami Ichikawa <masami.ichikawa@miraclelinux.com>