script: Add repository settings
Debian has entered the ELTS phase, the DEBIAN_MIRROR URL defined in Poky will
eventually no longer be available.
So, add setting to local.conf for refer to 'archive.debian.org'.

And, same to definition of DEBIAN_SECURITY_UPDATE_MIRROR, but 'debian-security'
archive does not yet exist, add it to MIRRORS for the future.