support: bmap-tools: change fetch URL
The original bmap-tools recipe refers to the intel repository on github.
The branch name in this repository has been changed form "master" to
"main" recently. So, bitbake cannot fetch it.

In addtion, the Intel repository is no longer supported.

Therefore, changed the URL to the active repository maintained by Yocto
Project.

Signed-off-by: Hirotaka Motai <hirotaka.motai@miraclelinux.com>